promote Snapshot interface
