+++ packages/sanddance-app/src/types.ts
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
-import { DataFile, SandDance, Snapshot } from '@msrvida/sanddance-explorer';
+import { DataFile, SandDance } from '@msrvida/sanddance-explorer';
 
 export type DataSourceType = 'sample' | 'local' | 'url';
 
@@ -13,6 +13,6 @@ export interface InsightMap {
   [id: string]: SandDance.specs.Insight;
 }
 
-export interface DataSourceSnapshot extends Snapshot {
+export interface DataSourceSnapshot extends SandDance.types.Snapshot {
   dataSource: DataSource;
 }
